dxd-e.c: make the pilot signal start at a unique point in its cycle on every file
- add option "--no-random" to preserve current behavior for testing purposes
- switch to a portable source of randomness rather than Linux getrandom()
- fix some printf()s argument type mismatches (fixes building for Windows)

+++ decode-dxd.c
@@ -73,8 +73,8 @@ int main (int argc, char **argv)
 
     }
 
-    fprintf (stderr, "%ld total PCM samples, %ld total DSD samples, %ld were embedded (all chans, %.1f%%)\n",
-        total_pcm_samples, total_dsd_samples, decodeTotalEmbeddedSamples (decoder),
+    fprintf (stderr, "%lld total PCM samples, %lld total DSD samples, %lld were embedded (all chans, %.1f%%)\n",
+        (long long) total_pcm_samples, (long long) total_dsd_samples, (long long) decodeTotalEmbeddedSamples (decoder),
         decodeTotalEmbeddedSamples (decoder) * 100.0 / total_dsd_samples / nchans);
 
     transitionDSDdumpstats (stderr);
